Move vue to its own folder outide of atomic structure
