+++ source/default/_patterns/02-molecules/vue-widget/index.js
@@ -13,14 +13,11 @@ import './_vue-cryptos.twig';
 import './_vue-cards.twig';
 import './_vue-clock.twig';
 
-// Module styles
-import './_vue-widget.scss';
-
 // All Vue widgets only execute on existence of DOM
-import FacetTableVue from './src/vue-cryptos';
-import FacetTableVueX from './src/vuex-cryptos';
-import VueCards from './src/vue-cards';
-import VueClock from './src/vue-clock';
+import FacetTableVue from '../../../clientside/vue-widget/src/vue-cryptos';
+import FacetTableVueX from '../../../clientside/vue-widget/src/vuex-cryptos';
+import VueCards from '../../../clientside/vue-widget/src/vue-cards';
+import VueClock from '../../../clientside/vue-widget/src/vue-clock';
 
 // Render Vue elements as soon as possible
 if (document.getElementById('vue-cryptos')) {
