+++ source/default/clientside/vue-widget/__tests__/mock-data.json
