+++ source/default/clientside/vue-widget/__tests__/vue-cryptos.test.js
