+++ source/default/clientside/vue-widget/src/vue-cards/app.vue
