+++ source/default/clientside/vue-widget/src/vue-cards/components/card.vue
