+++ source/default/clientside/vue-widget/src/vue-cards/index.js
