+++ source/default/clientside/vue-widget/src/vue-clock/index.js
