+++ source/default/clientside/vue-widget/src/vue-clock/vue-clock.vue
