+++ source/default/clientside/vue-widget/src/vue-cryptos/components/facet-table-facets.vue
