+++ source/default/clientside/vue-widget/src/vue-cryptos/facet-table.vue
