+++ source/default/clientside/vue-widget/src/vue-cryptos/index.js
