+++ source/default/clientside/vue-widget/src/vuex-cryptos/components/facet-table-facets.vue
