+++ source/default/clientside/vue-widget/src/vuex-cryptos/facet-table.vue
