+++ source/default/clientside/vue-widget/src/vuex-cryptos/index.js
