--- source/default/widgets/vue-widget/_vue-widget.scss
@@ -1,6 +1,0 @@
-// The non-printing variables, functions, and mixins of 00-protons/variables
-// are provided to this file via:
-//   webpack.particle.js: sass-loader.data|includePaths
-
-.vue-widget {
-}
